requirements.txt
moviepy==1.0.3
transformers==4.35.2 
torchaudio==2.7.1
librosa==0.10.1 
numpy==1.26.0 
pandas==2.1.0 
matplotlib==3.8.0 
seaborn==0.13.0
openai-whisper==20240930
gradio==5.23.1
pyannote.audio==3.3.2
ffmpeg-python==1.4.0